adding the ssl_model jupyter nb
The code first creates the feature matrix and target array by transforming the variables. Once we have the feature and target (for different thresholds t=21 and 30) then we model it using logit, KNN(for multiple ks), naive bayes, decision tree and random forest. 
we also generate ROCs for all the models.